creating InputRadio and deprecating Radio
