+++ packages/react/ds/src/radio/types.ts
@@ -1,23 +1,6 @@
-import { InputHTMLAttributes } from 'react';
-import type { TextInputProps } from '../text-input/text-input.js';
-
-export const RadioSizeEnum = {
-  Large: 'lg',
-  Medium: 'md',
-  Small: 'sm',
-} as const;
-
-export type RadioSizeType = (typeof RadioSizeEnum)[keyof typeof RadioSizeEnum];
-
-export type RadioProps = {
-  label?: string;
-  hint?: string;
-  size?: RadioSizeType;
-  conditionalInput?: TextInputProps;
-} & Omit<InputHTMLAttributes<HTMLInputElement>, 'size'>;
-
-export type RadioGroupProps = {
-  groupId: string;
-  inline?: boolean;
-  onChange?: (event?: React.ChangeEvent<HTMLInputElement>) => void;
-};
+export type {
+  InputRadioSizeType as RadioSizeType,
+  InputRadioProps as RadioProps,
+  RadioGroupProps,
+} from '../input-radio/types.js';
+export { InputRadioSizeEnum as RadioSizeEnum } from '../input-radio/types.js';
