+++ packages/react/ds/src/text-input/text-input.tsx
@@ -1,12 +1,12 @@
 'use client';
 
 import { InputText } from '../input-text/input-text.js';
-import type { InputProps } from '../input-text/type.js';
+import type { InputTextProps } from '../input-text/type.js';
 
 /**
  * @deprecated Use `<Input />` instead of `<TextInput />`.
  */
 export const TextInput = InputText;
-export type TextInputProps = InputProps;
+export type TextInputProps = InputTextProps;
 
 TextInput.displayName = 'TextInput';
